WEP-10: update tests structure
