+++ backend/tests/api/test_ping.py
@@ -1,0 +1,5 @@
+def test_ping(client):
+    response = client.get('/ping')
+
+    assert response.status_code == 200
+    assert response.text == 'pong'
